Rollup refactoring
